+++ src/lib/layout/mobile/utils.ts
@@ -34,7 +34,7 @@ export const getNavDrawerLite = (
             },
             {
               name: "Query",
-              slug: "/query",
+              slug: "/interact-contract",
               icon: "query" as IconKeys,
             },
           ]
@@ -117,6 +117,11 @@ export const getNavDrawerFull = (
               name: "0x1 Page",
               slug: "/account/0x1",
               icon: "hex" as IconKeys,
+            },
+            {
+              name: "View Module",
+              slug: "/interact",
+              icon: "query" as IconKeys,
             },
           ]
         : []),