+++ src/lib/pages/interact/component/drawer/selector/ModuleSelectorDisplay.tsx
@@ -1,7 +1,8 @@
-import { Button, Flex } from "@chakra-ui/react";
+import { Button, Flex, IconButton, Text } from "@chakra-ui/react";
 import type { Dispatch, SetStateAction } from "react";
 
 import type { DisplayMode, SelectedAddress } from "../types";
+import { useMobile } from "lib/app-provider";
 import { CopyLink } from "lib/components/CopyLink";
 import { CustomIcon } from "lib/components/icon";
 import { LabelText } from "lib/components/LabelText";
@@ -14,43 +15,81 @@ interface ModuleSelectorDisplayProps {
 export const ModuleSelectorDisplay = ({
   selectedAddress,
   setMode,
-}: ModuleSelectorDisplayProps) => (
-  <Flex
-    className="selector-display"
-    justifyContent="space-between"
-    w="full"
-    bgColor="gray.800"
-    px={4}
-    py={3}
-    mb={6}
-    borderRadius={8}
-    alignItems="center"
-  >
-    <Flex flex={1}>
-      <LabelText label="Viewing Address">
-        <CopyLink
-          value={selectedAddress.address}
-          type="user_address"
-          showCopyOnHover
-        />
-      </LabelText>
-    </Flex>
-    <Flex flex={1}>
-      <LabelText label="Hex">
-        <CopyLink
-          value={selectedAddress.hex}
-          type="user_address"
-          showCopyOnHover
-        />
-      </LabelText>
-    </Flex>
-    <Button
-      variant="outline-white"
-      size="sm"
-      leftIcon={<CustomIcon name="swap" boxSize={3} />}
-      onClick={() => setMode("input")}
+}: ModuleSelectorDisplayProps) => {
+  const isMobile = useMobile();
+  return (
+    <Flex
+      className="selector-display"
+      justifyContent="space-between"
+      w="full"
+      bgColor="gray.800"
+      px={4}
+      py={3}
+      mb={6}
+      borderRadius={8}
+      alignItems="center"
     >
-      Change Address
-    </Button>
-  </Flex>
-);
+      {isMobile ? (
+        <>
+          <Flex direction="column" gap={1}>
+            <Flex alignItems="center" gap={2}>
+              <Text variant="body2" color="text.dark">
+                VM Address:
+              </Text>
+              <CopyLink
+                value={selectedAddress.address}
+                type="user_address"
+                isTruncate
+              />
+            </Flex>
+            <Flex alignItems="center" gap={2}>
+              <Text variant="body2" color="text.dark">
+                HEX Address:
+              </Text>
+              <CopyLink
+                value={selectedAddress.hex}
+                type="user_address"
+                isTruncate
+              />
+            </Flex>
+          </Flex>
+          <IconButton
+            variant="outline-white"
+            onClick={() => setMode("input")}
+            icon={<CustomIcon name="swap" boxSize={3} />}
+            aria-label="reattach schema"
+          />
+        </>
+      ) : (
+        <>
+          <Flex flex={1}>
+            <LabelText label="Viewing Address">
+              <CopyLink
+                value={selectedAddress.address}
+                type="user_address"
+                showCopyOnHover
+              />
+            </LabelText>
+          </Flex>
+          <Flex flex={1}>
+            <LabelText label="Hex">
+              <CopyLink
+                value={selectedAddress.hex}
+                type="user_address"
+                showCopyOnHover
+              />
+            </LabelText>
+          </Flex>
+          <Button
+            variant="outline-white"
+            size="sm"
+            leftIcon={<CustomIcon name="swap" boxSize={3} />}
+            onClick={() => setMode("input")}
+          >
+            Change Address
+          </Button>
+        </>
+      )}
+    </Flex>
+  );
+};