+++ src/lib/services/account/index.ts
@@ -1,5 +1,6 @@
 import type { UseQueryOptions, UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
+import { useCallback } from "react";
 
 import type { AccountData, AccountTableCounts } from "../types";
 import {
@@ -10,18 +11,18 @@ import {
   useTierConfig,
   useWasmConfig,
 } from "lib/app-provider";
-import type { BechAddr } from "lib/types";
+import type { AccountType, BechAddr, Option } from "lib/types";
 
 import { getAccountData, getAccountTableCounts } from "./api";
-import { getAccountDataLcd } from "./lcd";
+import { getAccountDataLcd, getAccountTypeLcd } from "./lcd";
 
 export const useAccountData = (
   address: BechAddr
 ): UseQueryResult<AccountData> => {
   const isFullTier = useTierConfig() === "full";
-  const lcdEndpoint = useLcdEndpoint();
   const apiEndpoint = useBaseApiRoute("accounts");
-  const endpoint = isFullTier ? lcdEndpoint : apiEndpoint;
+  const lcdEndpoint = useLcdEndpoint();
+  const endpoint = isFullTier ? apiEndpoint : lcdEndpoint;
 
   return useQuery(
     [CELATONE_QUERY_KEYS.ACCOUNT_DATA, endpoint, address],
@@ -51,6 +52,35 @@ export const useAccountTableCounts = (
     ],
     async () => getAccountTableCounts(endpoint, address, isGov, isWasm),
     { retry: 1, refetchOnWindowFocus: false, ...options }
+  );
+};
+
+export const useAccountTypeLcd = (
+  address: Option<BechAddr>,
+  options: Pick<
+    UseQueryOptions<AccountType, Error>,
+    "enabled" | "onSuccess" | "onError"
+  > = {}
+): UseQueryResult<AccountType> => {
+  const lcdEndpoint = useLcdEndpoint();
+
+  const queryFn = useCallback(async () => {
+    if (!address)
+      throw new Error(
+        "Fetching account type: failed to retrieve address. (useAccountTypeLcd)"
+      );
+
+    return getAccountTypeLcd(lcdEndpoint, address);
+  }, [lcdEndpoint, address]);
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.ACCOUNT_TYPE_LCD, lcdEndpoint, address],
+    queryFn,
+    {
+      ...options,
+      retry: 1,
+      refetchOnWindowFocus: false,
+    }
   );
 };