+++ src/lib/services/searchService.ts
@@ -18,7 +18,7 @@ import {
   splitModule,
 } from "lib/utils";
 
-import { useBlockData, useBlockDataLcd } from "./block";
+import { useBlockData } from "./block";
 import { useModuleByAddressLcd } from "./move/module";
 import { useAddressByIcnsNameLcd, useIcnsNamesByAddressLcd } from "./name";
 import { usePoolByPoolId } from "./poolService";
@@ -164,13 +164,11 @@ export const useSearchHandler = (
     Number(debouncedKeyword),
     isPosDecimal(debouncedKeyword) && isFullTier
   );
-  const blockLcd = useBlockDataLcd(
-    Number(debouncedKeyword),
-    isPosDecimal(debouncedKeyword) && !isFullTier
-  );
-  const { data: blockData, isFetching: blockFetching } = isFullTier
-    ? blockApi
-    : blockLcd;
+  const { foundBlock, isFetching: blockFetching } = useMemo(() => {
+    if (isPosDecimal(debouncedKeyword) && !isFullTier)
+      return { foundBlock: true, isFetching: false };
+    return { foundBlock: blockApi.data, isFetching: blockApi.isFetching };
+  }, [blockApi.data, blockApi.isFetching, debouncedKeyword, isFullTier]);
 
   // Proposal
   const { data: proposalApiData, isFetching: proposalApiIsFetching } =
@@ -254,7 +252,7 @@ export const useSearchHandler = (
       moduleData && "Module Path",
       txData && "Transaction Hash",
       codeData && "Code ID",
-      blockData && "Block",
+      foundBlock && "Block",
       proposalData?.info && "Proposal ID",
       validatorData && "Validator Address",
       poolData && "Pool ID",